Added new ways to find canonical urls and fixed bugs
The bot can now find more and even second-level canonical URLs. This means that the bot is now more capable than ever. I've changed a lot of methods to improve the readability and maintainability. Also, to reduce errors, the bot now uses mobile chrome user agents. And I fixed a couple of bugs. Overall, this update improves the trustworthiness or AmputatorBot.